requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
lightgbm>=4.0
mlflow>=2.13
# PySpark + DeltaLake (optional for local; preinstalled on Databricks)
pyspark>=3.4
delta-spark>=3.2.0
pyarrow>=14.0
